refactor: выполнить рефакторинг всех сервисов мэйн модуля.
