+++ ewm-service/src/main/java/ru/practicum/ewmservice/compilation/controller/CompilationPublicController.java
@@ -1,0 +1,30 @@
+package ru.practicum.ewmservice.compilation.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.ewmservice.compilation.dto.CompilationDto;
+import ru.practicum.ewmservice.compilation.service.CompilationService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping(path = "/compilations")
+@RequiredArgsConstructor(onConstructor_ = @Autowired)
+@Validated
+public class CompilationPublicController {
+    private final CompilationService compilationService;
+
+    @GetMapping
+    public List<CompilationDto> getAll(@RequestParam(name = "pinned") boolean pinned,
+                                       @RequestParam(name = "from", defaultValue = "0") int from,
+                                       @RequestParam(name = "size", defaultValue = "10") int size) {
+        return compilationService.getAll(pinned, from, size);
+    }
+
+    @GetMapping(path = "/{compId}")
+    public CompilationDto getById(@PathVariable("compId") long compId) {
+        return compilationService.getById(compId);
+    }
+}
