+++ ewm-service/src/main/java/ru/practicum/ewmservice/compilation/model/Compilation.java
@@ -1,0 +1,31 @@
+package ru.practicum.ewmservice.compilation.model;
+
+import lombok.Getter;
+import lombok.Setter;
+import ru.practicum.ewmservice.event.model.Event;
+
+import javax.persistence.*;
+import java.util.Set;
+
+@Entity
+@Table(name = "compilations")
+@Setter
+@Getter
+public class Compilation {
+    @Id
+    @Column(name = "compilation_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    @JoinTable(
+            name = "compilations_events",
+            joinColumns = @JoinColumn(name = "compilation_id"),
+            inverseJoinColumns = @JoinColumn(name = "event_id")
+    )
+    private Set<Event> events;
+    @Column(name = "pinned")
+    private boolean pinned;
+    @Column(name = "title")
+    private String title;
+
+}
