+++ ewm-service/src/main/java/ru/practicum/ewmservice/compilation/storage/CompilationRepo.java
@@ -1,0 +1,13 @@
+package ru.practicum.ewmservice.compilation.storage;
+
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.practicum.ewmservice.compilation.model.Compilation;
+
+import java.util.List;
+
+@Repository
+public interface CompilationRepo extends JpaRepository<Compilation, Long> {
+    List<Compilation> findAllByPinned(boolean pinned, Pageable pageable);
+}
