+++ ewm-service/src/main/java/ru/practicum/ewmservice/event/controller/EventPublicController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import ru.practicum.ewmservice.event.dto.EventFullDto;
 import ru.practicum.ewmservice.event.dto.EventShortDto;
 import ru.practicum.ewmservice.event.dto.EventSorts;
-import ru.practicum.ewmservice.event.service.EventService;
+import ru.practicum.ewmservice.event.service.EventPublicService;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -17,7 +17,7 @@ import java.util.List;
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
 @Validated
 public class EventPublicController {
-    private final EventService eventService;
+    private final EventPublicService eventPublicService;
 
     @GetMapping
     public List<EventShortDto> findByFilter(@RequestParam(name = "text") String text,
@@ -32,11 +32,11 @@ public class EventPublicController {
                                             @RequestParam(name = "sort")EventSorts sort, // todo сделай конвертер
                                             @RequestParam(name = "from", defaultValue = "0") int from,
                                             @RequestParam(name = "size", defaultValue = "10") int size) {
-        return eventService.getAllPublic(text, categories, paid, rangeStart, rangeEnd, onlyAvailable, sort, from, size);
+        return eventPublicService.getAll(text, categories, paid, rangeStart, rangeEnd, onlyAvailable, sort, from, size);
     }
 
     @GetMapping("/{id}")
     public EventFullDto getById(@PathVariable("id") long eventId) {
-        return eventService.getPublicById(eventId);
+        return eventPublicService.getById(eventId);
     }
 }
