+++ ewm-service/src/main/java/ru/practicum/ewmservice/participation_request/model/EventRequest.java
@@ -3,7 +3,7 @@ package ru.practicum.ewmservice.participation_request.model;
 import lombok.Getter;
 import lombok.Setter;
 import ru.practicum.ewmservice.event.model.Event;
-import ru.practicum.ewmservice.users.model.User;
+import ru.practicum.ewmservice.user.model.User;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
