+++ ewm-service/src/main/java/ru/practicum/ewmservice/participation_request/storage/EventRequestRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import ru.practicum.ewmservice.event.model.Event;
 import ru.practicum.ewmservice.participation_request.model.EventRequest;
-import ru.practicum.ewmservice.users.model.User;
+import ru.practicum.ewmservice.user.model.User;
 
 import java.util.List;
