+++ ewm-service/src/main/java/ru/practicum/ewmservice/user/dto/UserDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewmservice.users.dto;
+package ru.practicum.ewmservice.user.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +7,8 @@ import lombok.RequiredArgsConstructor;
 @Builder(toBuilder = true)
 @Getter
 @RequiredArgsConstructor
-public class UserShortDto {
+public class UserDto {
     private final Long id;
     private final String name;
+    private final String email;
 }
