+++ ewm-service/src/main/java/ru/practicum/ewmservice/user/dto/UserShortDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewmservice.users.dto;
+package ru.practicum.ewmservice.user.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,8 +7,7 @@ import lombok.RequiredArgsConstructor;
 @Builder(toBuilder = true)
 @Getter
 @RequiredArgsConstructor
-public class UserDto {
+public class UserShortDto {
     private final Long id;
     private final String name;
-    private final String email;
 }
