+++ ewm-service/src/main/java/ru/practicum/ewmservice/user/model/UserRoles.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewmservice.users.model;
+package ru.practicum.ewmservice.user.model;
 
 public enum UserRoles { // todo удалить или реализовать
     ADMIN,
