+++ ewm-service/src/main/java/ru/practicum/ewmservice/user/storage/UserRepo.java
@@ -1,11 +1,11 @@
-package ru.practicum.ewmservice.users.storage;
+package ru.practicum.ewmservice.user.storage;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import ru.practicum.ewmservice.users.model.User;
+import ru.practicum.ewmservice.user.model.User;
 
 import java.util.List;
