+++ ewm-service/src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 DROP TABLE IF EXISTS users, categories, event_states, locations, events,
-    event_request_stats, event_requests;
+    event_request_stats, event_requests, compilations, compilations_events;
 
 CREATE table IF NOT EXISTS users(
     user_id bigint generated by default as identity primary key,
@@ -56,7 +56,19 @@ CREATE table IF NOT EXISTS event_requests(
     event_id  bigint REFERENCES events (event_id) ON DELETE CASCADE,
     requester_id bigint REFERENCES users (user_id) ON DELETE CASCADE,
     status_id bigint REFERENCES event_request_stats (stat_id) ON DELETE CASCADE
-)
+);
+
+CREATE table IF NOT EXISTS compilations(
+    compilation_id bigint generated by default as identity primary key,
+    pinned bool NOT NULL,
+    title varchar NOT NULL
+);
+
+CREATE table IF NOT EXISTS compilations_events(
+    compilation_id bigint REFERENCES compilations (compilation_id) ON DELETE CASCADE,
+    event_id bigint REFERENCES events (event_id) ON DELETE CASCADE,
+    PRIMARY KEY (compilation_id, event_id)
+);
